rename: 프로필 관련 컴포넌트를 한 파일에 옮김
