+++ src/components/header/profile/UserProfileRead.tsx
@@ -8,7 +8,7 @@ import { GrFormNext } from 'react-icons/gr';
 import { IoClose } from 'react-icons/io5';
 
 import styles from './UserProfileRead.module.css';
-import Logout from '../auth/LogoutButton';
+import Logout from '../../auth/LogoutButton';
 
 export type UserProfileData = { name: string; profile_url: string | null };
